Разобралсиь с контроллерами, ща всё будет

+++ src/main/java/com/example/demo/service/impl/DriverServiceImpl.java
@@ -38,6 +38,7 @@ public class DriverServiceImpl implements DriverService {
 
     @Override
     public void deleteDriver(Long id) {
-        driversRepository.deleteById(id);
+        if(driversRepository.existsById(id))
+            driversRepository.deleteById(id);
     }
 }
